### 2019-11-14 #1892 @whedon
[PRE REVIEW]: CapacityExpansion: A capacity expansion modeling framework in Julia
**Submitting author:** @YoungFaithful (<a href="http://orcid.org/0000-0002-1992-310X">Lucas Elias Kuepper</a>)
**Repository:** <a href="https://github.com/YoungFaithful/CapacityExpansion.jl/tree/paper" target ="_blank">https://github.com/YoungFaithful/CapacityExpansion.jl/tree/paper</a>
**Version:** v0.2.0
