### 2019-11-14 #1892 @labarba
I think maybe the authors provided the wrong repository URL. They provided `https://github.com/YoungFaithful/CapacityExpansion.jl/tree/paper`.

I will change it to `https://github.com/YoungFaithful/CapacityExpansion.jl/tree/paper` and try to re-compile.
