### 2019-11-14 #1892 @labarba
@YoungFaithful — Where is your JOSS paper? I poked around in the repo but could not find it.
