### 2019-11-14 #1892 @arfon
> @YoungFaithful — Where is your JOSS paper? I poked around in the repo but could not find it.

Looks like it's in the `paper` branch @labarba.
