### 2019-11-14 #1892 @labarba
👋  @katyhuff — Could you handle this submission as editor? It is in the area of energy systems management.
