### 2019-11-28 7f5661c
Update examples and documentation to split settings (OptConfig) and the optimization (run_opt)
