+++ src/optim_problems/run_opt.jl
@@ -1,6 +1,6 @@
 """
     run_opt(ts_data::ClustData,opt_data::OptDataCEP,config::OptConfig)
-Organizing the actual setup and run of the CEP-Problem. This function shouldn't be called by a user, but from within the other `run_opt`-functions
+Organizing the actual setup and run of the CEP-Problem.
 Required elements are:
 - `ts_data`: The time-series data.
 - `opt_data`: In this case the OptDataCEP that contains information on costs, nodes, techs and for transmission also on lines.
@@ -83,8 +83,11 @@ end
 
 """
      run_opt(ts_data::ClustData,opt_data::OptDataCEP,config::OptConfig,fixed_design_variables::Dict{String,Any},optimizer::DataTyple;lost_el_load_cost::Number=Inf,lost_CO2_emission_cost::Number)
-This problem runs the operational optimization problem only, with fixed design variables.
-provide the fixed design variables and the `config` of the previous step (design run or another opterational run)
+This problem runs the operational optimization problem only, with fixed design variables. Provide the fixed design variables and the `config` of the previous step (design run or another opterational run). We recommend to alternatively run OptConfig and the optimization in two steps:
+```julia
+opt_config = OptConfig(old_config,fixed_design_variables; lost_load_cost=lost_load_cost, lost_emission_cost=lost_emission_cost)
+run_opt(ts_data, opt_data, opt_config)
+```
 Required elements are:
 - `ts_data`: The time-series data, which should be be the original time-series data for this operational run. The `keys(ts_data.data)` need to match the `[time_series_name]-[node]`
 - `opt_data`: In this case the OptDataCEP that contains information on costs, nodes, techs and for transmission also on lines. - Should be the same as in the design run.
@@ -125,36 +128,40 @@ end
            scale::Dict{Symbol,Int}=Dict{Symbol,Int}(:COST => 1e9, :CAP => 1e3, :GEN => 1e3, :SLACK => 1e3, :INTRASTOR => 1e3, :INTERSTOR => 1e6, :FLOW => 1e3, :TRANS =>1e3, :LL => 1e6, :LE => 1e9),
            print_flag::Bool=true,
            optimizer_config::Dict{Symbol,Any}=Dict{Symbol,Any}(),
-           round_sigdigits::Int=9)
-Wrapper function for type of optimization problem for the CEP-Problem (NOTE: identifier is the type of `opt_data` - in this case OptDataCEP - so identification as CEP problem).
+           round_sigdigits::Int=9,
+           time_series_config::Dict{String,Any}=Dict{String,Any}())
+Wrapper function for type of optimization problem for the CEP-Problem (NOTE: identifier is the type of `opt_data` - in this case OptDataCEP - so identification as CEP problem). We recommend to alternatively run OptConfig and the optimization in two steps to get more information about the kwargs:
+```julia
+opt_config = OptConfig(ts_data, opt_data, optimizer; [options to tweak the model] )
+run_opt(ts_data, opt_data, opt_config)
+```
 Required elements are:
 - `ts_data`: The time-series data, which could either be the original input data or some aggregated time-series data. The `keys(ts_data.data)` need to match the `[time_series_name]-[node]`
 - `opt_data`: The OptDataCEP that contains information on costs, nodes, techs and for transmission also on lines.
 - `optimizer`: The used optimizer, which could e.g. be Clp: `using Clp` `optimizer=Clp.Optimizer` or Gurobi: `using Gurobi` `optimizer=Gurobi.Optimizer`.
-Options to tweak the model are:
-- `descriptor`: A name for the model
+Options to tweak the model are used to configure the `OptConfig`:
+- `descriptor`: String with the name of this paricular model like "kmeans-10-co2-500"
[…]
