+++ src/utils/datastructs.jl
@@ -38,15 +38,20 @@ struct OptVariable{T,N,Ax,L<:NTuple{N,Dict}} <: AbstractArray{T,N}
 end
 
 """
-  OptConfig{tech::Dict{String,Bool}
+  OptConfig{descriptor::String
+      region::String
       model::Dict{String,Bool}
-      limit_emission::Dict{String,Number}
+      limit_emission::Dict{String,Dict{String,Number}}
+      infrastructure::Dict{String,Array}
       scale::Dict{Symbol,Int}
-      print_flag::Bool
       optimizer::DataType
       optimizer_config::Dict{Symbol,Any}
-      round_sigdigits::Int
-      time_series_config::OptConfig}
+      time_series::Dict{String,Any}
+      fixed_design_variables::Dict{String,Any}
+      lost_load_cost::Dict{String,Number}
+      lost_emission_cost::Dict{String,Number}
+      print_flag::Bool
+      round_sigdigits::Int}
 
 contains the information that tweaks the model
 """
@@ -280,44 +285,44 @@ end
 
 """
         OptConfig(ts_data::ClustData,
-                    opt_data::OptDataCEP;
-                    descriptor::String="",
-                    storage_type::String="none",
-                    limit_emission::Dict{String,Number}=Dict{String,Number}(),
-                    demand::Bool=true,
-                    dispatchable_generation::Bool=true,
-                    non_dispatchable_generation::Bool=true,
-                    conversion::Bool=false,
-                    transmission::Bool=false,
-                    lost_load_cost::Dict{String,Number}=Dict{String,Number}(),
-                    lost_emission_cost::Dict{String,Number}=Dict{String,Number}(),
-                    limit_emission::Dict{String,Number}=Dict{String,Number}(),
-                    infrastructure::Dict{String,Array}=Dict{String,Array}("existing"=>["demand"],"limit"=>Array{String,1}()),
-                    scale::Dict{Symbol,Int}=Dict{Symbol,Int}(:COST => 1e9, :CAP => 1e3, :GEN => 1e3, :SLACK => 1e3, :INTRASTOR => 1e3, :INTERSTOR => 1e6, :FLOW => 1e3, :TRANS =>1e3, :LL => 1e6, :LE => 1e9),
-                    print_flag::Bool=true,
-                    optimizer_config::Dict{Symbol,Any}=Dict{Symbol,Any}(),
-                    round_sigdigits::Int=9,
-                    time_series::Dict{String,Any}=Dict{String,Any}())
+                opt_data::OptDataCEP,
+                optimizer::DataType;
+                descriptor::String="",
+                storage_type::String="none",
+                demand::Bool=true,
+                dispatchable_generation::Bool=true,
+                non_dispatchable_generation::Bool=true,
+                conversion::Bool=false,
+                transmission::Bool=false,
+                lost_load_cost::Dict{String,Number}=Dict{String,Number}(),
+                lost_emission_cost::Dict{String,Number}=Dict{String,Number}(),
+                limit_emission::Dict{String,Number}=Dict{String,Number}(),
+                infrastructure::Dict{String,Array}=Dict{String,Array}("existing"=>["demand"],"limit"=>Array{String,1}()),
+                scale::Dict{Symbol,Int}=Dict{Symbol,Int}(:COST => 1e9, :CAP => 1e3, :GEN => 1e3, :SLACK => 1e3, :INTRASTOR => 1e3, :INTERSTOR => 1e6, :FLOW => 1e3, :TRANS =>1e3, :LL => 1e6, :LE => 1e9),
+                print_flag::Bool=true,
+                optimizer_config::Dict{Symbol,Any}=Dict{Symbol,Any}(),
[…]
