### 2019-11-29 1e76d56
new asset formulation Documenter 0.24
