+++ docs/make.jl
@@ -13,16 +13,16 @@ makedocs(sitename="CapacityExpansion.jl",
         "Teaching" => ["teaching.md", "homework.md"],
         ],
     format = Documenter.HTML(assets = [
-        "assets/cep_text.svg",
-        "assets/opt_cep.svg",
-        "assets/workflow.svg",
-        "assets/GER_1.svg",
-        "assets/GER_18.svg",
-        "assets/CA_1.svg",
-        "assets/CA_14.svg",
-        "assets/opt_cep_cap_plot",
-        "assets/preparing_clust_data_load",
-        "assets/preparing_clust_data_agg"])
+        asset("assets/cep_text.svg", class=:ico, islocal=true),
+        asset("assets/opt_cep.svg", class=:ico, islocal=true),
+        asset("assets/workflow.svg", class=:ico, islocal=true),
+        asset("assets/GER_1.svg", class=:ico, islocal=true),
+        asset("assets/GER_18.svg", class=:ico, islocal=true),
+        asset("assets/CA_1.svg", class=:ico, islocal=true),
+        asset("assets/CA_14.svg", class=:ico, islocal=true),
+        asset("assets/opt_cep_cap_plot.svg", class=:ico, islocal=true),
+        asset("assets/preparing_clust_data_load.svg", class=:ico, islocal=true),
+        asset("assets/preparing_clust_data_agg.svg", class=:ico, islocal=true)])
         )
 
 deploydocs(repo = "github.com/YoungFaithful/CapacityExpansion.jl.git", devbranch="dev")
