### 2020-01-21 #2034 @whedon
[REVIEW]: CapacityExpansion: A capacity expansion modeling framework in Julia
**Submitting author:** @YoungFaithful (<a href="http://orcid.org/0000-0002-1992-310X">Lucas Elias Kuepper</a>)
**Repository:** <a href="https://github.com/YoungFaithful/CapacityExpansion.jl" target ="_blank">https://github.com/YoungFaithful/CapacityExpansion.jl</a>
**Version:** v0.2.0
