### 2020-02-05 #2034 @YoungFaithful
@arfon - Do you have a clue how to generate the pdf? In comparison to the pre-review, the Repository now contains `/tree/paper` and neither `@whedon generate pdf` nor `@whedon generate pdf from branch paper` worked
