### 2020-02-05 #2034 @arfon
@YoungFaithful - this is because the URL for the repository needs to be something Whedon can clone (e.g. `https://github.com/YoungFaithful/CapacityExpansion.jl/`).

I've now fixed this.
