### 2020-02-05 9e202e9
add fixed battery size and fixed transmission data
