### 2020-02-14 #2034 @sjpfenninger
All right, I'm through with my review. Generally looks good, but I do have a couple of bigger points below. Happy to clarify anything if unclear!

## Minor comments

* One of the paper authors seems to have made no contributions to the code.

* Installation: is using environments generally a thing in the Julia world? Would it make sense to update instructions accordingly? (https://docs.julialang.org/en/v1/stdlib/Pkg/index.html)

* Documentation was hard to find. Link more prominently in the README, ideally with a separate section that contains the link(s) to where built docs live? It took me a while to spot that the way to get to docs is through the badge at the top of the README. Also set the repository link to point to the docs.

* I cannot see contribution or community guidelines anywhere.

* There are quite a few spelling and grammar errors throughout the documentation, making it harder than necessary to parse. Not necessarily for now, but at some point, a round of proofreading might be an hour or two well spent.

* In the paper: "The scale of the modeled energy system can range from quarters to entire global regions" - I believe you mean "districts", not "quarters".

## Statement of need in documentation and software paper / state of field

I think these need some work.

You describe the software as "Input-data-scaling, multi-sector capacity expansion modeling framework." Can you clarify what "input-data-scaling" means? Also, as far as I see, what you call "capacity expansion modeling" is functionally equivalent to the term "energy system modeling", correct?

It would be useful here to give more information on how, if at all, the functionality of your tool differs from existing software packages in this area (especially since you already list a selection of them, and since there are so many pre-existing packages which all essentially do the same thing). Is there any reason for a prospective user to go for your software rather than an older, more established, better tested tool?

You should also list [PyPSA](https://github.com/PyPSA/PyPSA) amongst the existing packages.

You might also want to refer to EnergyModels.jl, which as far as I'm aware of is not yet officially published, but I think under active development -- and based on the first public commit, is older than CapacityExpansion.jl:

https://github.com/PyPSA/EnergyModels.jl

## Tests

There seem to be no automated tests. A `test` dir exists but not referred to anywhere. It looks like a single test model. Can you clarify what testing is done?
