### 2020-02-27 #2034 @YoungFaithful
> All right, I'm through with my review. Generally looks good, but I do have a couple of bigger points below. Happy to clarify anything if unclear!
> ## Minor comments
>
>     * One of the paper authors seems to have made no contributions to the code.
Our professor didn't contribute to the code
>
>     * Installation: is using environments generally a thing in the Julia world? Would it make sense to update instructions accordingly? (https://docs.julialang.org/en/v1/stdlib/Pkg/index.html)
I fear that I don't fully understand your question about environments in the Julia context: Each package defines its dependencies in `Project.toml` (https://github.com/YoungFaithful/CapacityExpansion.jl/blob/master/Project.toml). The `Pkg`-Manager uses the dependencies of the different packages and tries to satisfy all of them. The user only uses `using Pkg
Pkg.add("CapacityExpansion")``, which is the same to `] add CapacityExpansion`
>
>     * Documentation was hard to find. Link more prominently in the README, ideally with a separate section that contains the link(s) to where built docs live? It took me a while to spot that the way to get to docs is through the badge at the top of the README. Also set the repository link to point to the docs.
Great point. Wasn't even aware of repository links :+1:
>
>     * I cannot see contribution or community guidelines anywhere.
Good catch that the README doesn't include a link to them.
>
>     * There are quite a few spelling and grammar errors throughout the documentation, making it harder than necessary to parse. Not necessarily for now, but at some point, a round of proofreading might be an hour or two well spent.
I'll have a second look at this
>
>     * In the paper: "The scale of the modeled energy system can range from quarters to entire global regions" - I believe you mean "districts", not "quarters".
True :)
>
>
> ## Statement of need in documentation and software paper / state of field
>
> I think these need some work.
>
> You describe the software as "Input-data-scaling, multi-sector capacity expansion modeling framework." Can you clarify what "input-data-scaling" means? Also, as far as I see, what you call "capacity expansion modeling" is functionally equivalent to the term "energy system modeling", correct?
I wanted to express that the scale of the energy system is dependent on the input data. Which formulation would you suggest?
I think that the terms `capacity expansion modeling` and `energy system modeling` are equivalent.
>
> It would be useful here to give more information on how, if at all, the functionality of your tool differs from existing software packages in this area (especially since you already list a selection of them, and since there are so many pre-existing packages which all essentially do the same thing). Is there any reason for a prospective user to go for your software rather than an older, more established, better tested tool?
I added a section at the end of the comparison to other packages.
>
> You should also list [PyPSA](https://github.com/PyPSA/PyPSA) amongst the existing packages.
>
> You might also want to refer to EnergyModels.jl, which as far as I'm aware of is not yet officially published, but I think under active development -- and based on the first public commit, is older than CapacityExpansion.jl:
Oh great. Didn't know this one.
>
> https://github.com/PyPSA/EnergyModels.jl
> ## Tests
>
> There seem to be no automated tests. A `test` dir exists but not referred to anywhere. It looks like a single test model. Can you clarify what testing is done?
Testing validates the results of multiple cases (different optimization settings) for five energy system models against the expected values. The TX model is validated against the capacity expansion planning model of James Merrick.
