+++ docs/src/README_TX_1.md
@@ -11,6 +11,6 @@ Implemented with PV-price 0.5 $/W
 ##Other
 Alternatively for price of 1.0$/W edit .csv files and replace costs with:
 - fix: 4.776E+3 $/MW  cap: 1.032E+6 $/MW
-Assuptions for transformation:
-- demand mulitiplied with 1.48
-- solar devided by 1000
+Assumptions for transformation:
+- demand multiplied with 1.48
+- solar divided by 1000
