### 2020-03-02 #2034 @sjpfenninger
@YoungFaithful thanks for the clarifications and updates!

Just four more comments:

* "input-data-scaling": perhaps you mean "scale-independent"? "Input-data-scaling" implies that some part of your code actively scales input data.

* Testing: You do of course have a CI job set up, so that was my bad! Still, I would add your explanation of the tests to the docs and/or the README to briefly flag what testing is done, tell users (=potential contributors) how to run the tests, or even better, provide something like a Makefile that does that.

* Environments: My question is whether it would make sense to update the installation documentation to explicitly include creating a separate environment (i.e. something like `pkg> activate capacityexpansion` as a first installation step). See https://julialang.github.io/Pkg.jl/v1/environments/. As far as I understand the Julia package manager doesn't sequester all of the package's dependencies into an environment by default -- but I may be wrong about that.

* Thanks for adding the paragraph on what the added value of CapacityExpansion.jl is. I accept that it is the first published Julia package to do the things it does, but it is not the only such tool to permit policy constraints. See https://github.com/YoungFaithful/CapacityExpansion.jl/commit/151a9a986c7ccdac0602a9a307c8f46f95520026#r37583878.
