+++ docs/make.jl
@@ -11,6 +11,7 @@ makedocs(sitename="CapacityExpansion.jl",
         "Data" => ["preparing_clust_data.md", "preparing_opt_cep_data.md", "csv_structure.md", "README_GER_18.md", "README_GER_1.md", "README_CA_14.md", "README_CA_1.md", "README_TX_1.md"],
         "Optimization" => ["opt_cep.md","results_opt.md", "opt_cep_examples.md"],
         "Teaching" => ["teaching.md", "homework.md"],
+        "Testing" => "testing.md",
         ],
     format = Documenter.HTML(assets = [
         "assets/clust_for_opt_text.svg",
