### 2020-03-29 #2034 @YoungFaithful
@sjpfenninger thanks for your additional comments :)
> Just four more comments:
>
>     * "input-data-scaling": perhaps you mean "scale-independent"? "Input-data-scaling" implies that some part of your code actively scales input data.
That's a good way to phrase it. Thank you.
>
>     * Testing: You do of course have a CI job set up, so that was my bad! Still, I would add your explanation of the tests to the docs and/or the README to briefly flag what testing is done, tell users (=potential contributors) how to run the tests, or even better, provide something like a Makefile that does that.
I included a description of the testing in the README and added a new "testing" page in the documentation that includes a detailed step by step description on how to run the testing.
>
>     * Environments: My question is whether it would make sense to update the installation documentation to explicitly include creating a separate environment (i.e. something like `pkg> activate capacityexpansion` as a first installation step). See https://julialang.github.io/Pkg.jl/v1/environments/. As far as I understand the Julia package manager doesn't sequester all of the package's dependencies into an environment by default -- but I may be wrong about that.
I have a different understanding of the Pkg-Manager in that regard. I think that something like `pkg> activate capacityexpansion` is used in order to update the dependencies of the package, which should only be done by developers. julia only allows one environment to my best knowledge. All package dependencies have to be fulfilled in that environment and one cannot build something like containers for certain packages.
>
>     * Thanks for adding the paragraph on what the added value of CapacityExpansion.jl is. I accept that it is the first published Julia package to do the things it does, but it is not the only such tool to permit policy constraints. See [YoungFaithful/CapacityExpansion.jl@151a9a9#r37583878](https://github.com/YoungFaithful/CapacityExpansion.jl/commit/151a9a986c7ccdac0602a9a307c8f46f95520026#r37583878).
I agree that it is not the only package that allows to constrain policy constraints. To my best knowledge it is the only open source package that allows you to limit any kind of life cycle assessment category (also multiple at once) that you include in the input data. Maybe I am wrong about that. I excluded the sentence from the paper.
