### 2020-04-19 87e502d
Make documentation in README file more visible
