### 2020-06-08 a28a9b1
Fixed three small typos in README
