### 2020-06-08 YoungFaithful/CapacityExpansion.jl#55 @ali-ramadhan
Remove explicit list of dependencies in README?
A list of dependencies in the README might give beginners to Julia the impression that they must install these dependencies themselves as virtually every Julia package does not list dependencies in the README. They are always handled automatically.

Perhaps a more useful way of listing dependencies is to point out that they are listed in the `Project.toml` file, which is automatically updated every time the dependencies change while the README must be edited manually.

X-Ref: https://github.com/openjournals/joss-reviews/issues/2034
