### 2020-06-08 25c78e8
Fix two typos in documentation quickstart
