### 2020-06-08 YoungFaithful/CapacityExpansion.jl#57 @ali-ramadhan
Dead link to Julia in README
Here: https://github.com/YoungFaithful/CapacityExpansion.jl/blob/master/README.md#L7

I assume you meant to link to https://julialang.org/ or https://www.juliaopt.org/ ?

X-Ref: https://github.com/openjournals/joss-reviews/issues/2034
