### 2020-06-08 YoungFaithful/CapacityExpansion.jl#59 @ali-ramadhan
Travis documentation builds failing
This is the error I'm seeing from https://travis-ci.com/github/YoungFaithful/CapacityExpansion.jl/jobs/346302482

```julia
$ julia --project=docs/ docs/make.jl
ERROR: LoadError: Unrecognised asset class unknown for `assets/clust_for_opt_text.svg`
Stacktrace:
 [1] error(::String) at ./error.jl:33
 [2] Documenter.Writers.HTMLWriter.HTMLAsset(::Symbol, ::String, ::Bool) at /home/travis/.julia/packages/Documenter/PLD7m/src/Writers/HTMLWriter.jl:81
 [3] #3 at /home/travis/.julia/packages/Documenter/PLD7m/src/Writers/HTMLWriter.jl:348 [inlined]
 [4] iterate at ./generator.jl:47 [inlined]
 [5] _collect(::Array{String,1}, ::Base.Generator{Array{String,1},getfield(Documenter.Writers.HTMLWriter, Symbol("##3#5"))}, ::Base.EltypeUnknown, ::Base.HasShape{1}) at ./array.jl:632
 [6] map at ./array.jl:561 [inlined]
 [7] #HTML#2 at /home/travis/.julia/packages/Documenter/PLD7m/src/Writers/HTMLWriter.jl:346 [inlined]
 [8] (::getfield(Core, Symbol("#kw#Type")))(::NamedTuple{(:assets,),Tuple{Array{String,1}}}, ::Type{Documenter.Writers.HTMLWriter.HTML}) at ./none:0
 [9] top-level scope at none:0
 [10] include at ./boot.jl:317 [inlined]
 [11] include_relative(::Module, ::String) at ./loading.jl:1044
 [12] include(::Module, ::String) at ./sysimg.jl:29
 [13] exec_options(::Base.JLOptions) at ./client.jl:266
 [14] _start() at ./client.jl:425

in expression starting at /home/travis/build/YoungFaithful/CapacityExpansion.jl/docs/make.jl:5
```

X-Ref: openjournals/joss-reviews#2034
