+++ paper/paper.bib
@@ -71,7 +71,7 @@ author = {Weibezahn, Jens and Kendziorski, Mario},
 title = {Illustrating the Benefits of Openness: A Large-Scale Spatial Economic Dispatch Model Using the Julia Language},
 journal = {Energies},
 volume = {12},
-yera = {2019},
+year = {2019},
 number = {6},
 article-number = {1153},
 url = {https://www.mdpi.com/1996-1073/12/6/1153},
