### 2020-06-08 #2034 @ali-ramadhan
Hi @YoungFaithful, please accept my apologies for my extremely late review. I completely fell off the grid in the past few months. My review is below.

---

I think CapacityExpansion.jl is a well made and well documented Julia package. Everything worked out of the box and I was able to play around with it a bit. More detailed comments follow but I couldn't find any major faults. I opened a few issues and PRs on the repository but they're all minor. I'm happy to check off all the items on the review checklist but can't seem to do so right now.

## Functionality

> **Installation:** Does installation proceed as outlined in the documentation?

Yes. I was able to install CapacityExpansion 0.2.0 with no problems on Linux with Julia 1.4.2.

I only have one tiny suggestions:

The installation instructions use the older method of adding packages. It might be more future proof to suggest installing your package using the package mode available since Julia 1.0, e.g. `] add CapacityExpansion` which is also more concise. However, as you've pointed out earlier in this review they are identical and most Julia users are familiar with both methods.

> **Functionality:** Have the functional claims of the software been confirmed?

I believe so. I am not an expert in energy systems modeling but I was able to follow and play around with the example in the documentation quickstart which showed how to optimize the German renewable solar+wind energy grid subject to some constraints. The test suite seems to test these high-level functions pretty comprehensively.

> **Performance:** If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

No performance claims have been made.

## Documentation

> **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?

Yes. The problems solved by the software package are stated in the README and the software paper. The target audience is not explicitly stated, however the software package can be used to "address a diverse set of research questions" which to me implicitly states that the target audience is researchers. At the end of the software paper it states that it's been used as an educational tool as well, which is nice to see. So students are a potential target audience too.

> **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

Yes. However I might actually suggest removing the explicit list of dependencies from the README.

A list of dependencies in the README might give beginners to Julia the impression that they must install these dependencies themselves as virtually every Julia package does not list dependencies in the README. They are always handled automatically.

Perhaps a more useful way of listing dependencies is to point out that they are listed in the `Project.toml` file, which is automatically updated every time the dependencies change while the README must be edited manually.

I opened an issue about this: https://github.com/YoungFaithful/CapacityExpansion.jl/issues/55

> **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

Yes. I was able to run the example in the README. I was able to follow and run the example in the Quickstart. It is well documented to help users understand what the package is doing.

I even got back a result with the exact same objective so it seems very reproducible!

> **Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

Yes. The API is documented across a few different pages, one page per functionality which helps split things up. Examples, variable explanations, and plots are included which help a lot. Looks great!

> **Automated tests:** Are there automated tests or manual steps described so that the functionality of the software can be verified?

Automated tests are available and run through Travis CI Linux builds which is great. I was also able to run the tests on my laptop and they passed.

I have two small suggestions regarding the automated tests:

1. Rename the `test/cpl.jl` file to `test/runtests.jl` to follow Julia conventions. This way you can test the package automatically through the REPL. I opened a PR suggesting this change: https://github.com/YoungFaithful/CapacityExpansion.jl/pull/58

2. Currently tests only run on Julia 1.0 which is fine as Julia 1.0.5 is the long-term support version. However, a lot of people are running more recent Julia versions so it might be worth running Julia 1.4 or `dev` test builds to ensure the tests pass for those users.

> **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Yes.

## Software paper

> **Summary:** Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?

Yes.

> **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?

Yes.

> **State of the field:** Do the authors describe how this software compares to other commonly-used packages?

Yes. Other packages are extensively discussed which I found super useful.

> **Quality of writing:** Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?

It's great. I only found a few tiny typos which I opened a PR to fix: https://github.com/YoungFaithful/CapacityExpansion.jl/pull/60

> **References:** Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?

Yes. The citations were all appropriate and there is a good amount of them for completenes and pedagogical insight. I only found a tiny typo in one of the BiBTeX references and opened a PR to fix it: https://github.com/YoungFaithful/CapacityExpansion.jl/pull/61
