### 2020-06-13 5de739a
Documentation and minor corrections
