### 2020-06-13 f2fe4bb
Documentation (#62)
* paper draft

* Paper draft

* Update paper/paper.md

Co-Authored-By: Holger Teichgraeber <holgerteichgraeber@users.noreply.github.com>

* Update paper/paper.md

Co-Authored-By: Holger Teichgraeber <holgerteichgraeber@users.noreply.github.com>

* Update paper/paper.md

Co-Authored-By: Holger Teichgraeber <holgerteichgraeber@users.noreply.github.com>

* change requests added

* constraints, julia-ref

* Integrate anyMOD.jls

* Missing word included

* Review suggestions sjpfenninger

* Feedback sjpfenninger

* testing docs

* Fix tiny typos in `paper.md`

* Fix year for `We
[…]
