+++ docs/Project.toml
@@ -2,3 +2,7 @@
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 Clp = "e2554f3b-3117-50c0-817c-e040a3ddf72d"
+
+[compat]
+Documenter = "^0.24.0"
+Plots = "^0.27.0"

[…]
