### 2020-06-13 #2034 @YoungFaithful
Hi @ali-ramadhan,
thanks for your detailed review, the pull requests you prepared and issues you posted. I included your feedback and recommendations in a new update. Are you able to check off the items after whedon assigned you?
