### 2020-06-15 #2034 @ali-ramadhan
Thanks @YoungFaithful for addressing the issues I opened. I closed them as they've been resolved.

Still can't check the boxes in the review checklist. I'm logged into the correct account and tried clicking on the invitation link but got this message:

>  Sorry, we couldn't find that repository invitation. It is possible that the invitation was revoked or that you are not logged into the invited account.

so maybe the invitation expired or something? I had no issues ticking the boxes on a previous review but maybe a new invitation is needed for each review?
