### 2020-06-16 #2034 @YoungFaithful
@ali-ramadhan, thanks for your response and closing the issues. I have no idea about the issues with the boxes. @katyhuff, do you have an idea what we could do to activate the boxes?
