### 2020-06-16 #2034 @YoungFaithful
@sjpfenninger, I hope you're doing fine. I saw that there are some unchecked boxes remaining from your review. Did the last commits already solve the open aspects or is there anything else that we should improve?
