### 2020-06-16 #2034 @katyhuff
Sorry for the delay in responding.
@ali-ramadhan I'm not sure what's happened with your permissions, but I believe that it may be due to non-responsiveness earlier. I have checked all the boxes that you have affirmed as completed in your review. So, all are now checked except the confirmations of Conflict of interest and reading the JOSS code of conduct. Can you confirm @ali-ramadhan in a comment here the following statements :

1) I have read the JOSS conflict of interest (COI) policy and that: I have no COIs with reviewing this work or that any perceived COIs have been waived by JOSS for the purpose of this review.
Code of Conduct
2) I confirm that I read and will adhere to the JOSS code of conduct.
