### 2020-06-16 #2034 @ali-ramadhan
Thank you @katyhuff.

I have read the JOSS conflict of interest (COI) policy and that: I have no COIs with reviewing this work or that any perceived COIs have been waived by JOSS for the purpose of this review.

I confirm that I read and will adhere to the JOSS code of conduct.
