+++ CONTRIBUTING.md
@@ -1,4 +1,4 @@
 ## How to contribute to CapacityExpansion.jl
 Welcome! Thank you for considering to contribute to `CapacityExpansion.jl`. If you have a comment, question, feature request, or bug report, please open a new [issue](https://github.com/YoungFaithful/CapacityExpansion.jl/issues).
 
-If you like to file a bug report, or like to contribute to the documentation or the code (always welcome!), the [JuMP.jl Contributing.md](https://github.com/JuliaOpt/JuMP.jl/blob/master/CONTRIBUTING.md) has some great tips on how to get started. We follow the git workflow using a `dev`-branch for development and a `master`-branch for the stable versions.
+If you like to file a bug report or like to contribute to the documentation or the code (always welcome!), the [JuMP.jl Contributing.md](https://github.com/JuliaOpt/JuMP.jl/blob/master/CONTRIBUTING.md) has some great tips on how to get started. We follow the git workflow using a `dev`-branch for development and a `master`-branch for the stable versions.
