+++ v0.2.1/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"README_TX_1/#TX-1-1","page":"TX-1","title":"TX-1","text":"","category":"section"},{"location":"README_TX_1/#","page":"TX-1","title":"TX-1","text":"Single node of a power grid in Texas, no existing capacity","category":"page"},{"location":"README_TX_1/#","page":"TX-1","title":"TX-1","text":"Data from Merrick et al. 2016","category":"page"},{"location":"README_TX_1/#Cost-Data-1","page":"TX-1","title":"Cost Data","text":"","category":"section"},{"location":"README_TX_1/#","page":"TX-1","title":"TX-1","text":"Implemented with PV-price 0.5 /W","category":"page"},{"location":"README_TX_1/#","page":"TX-1","title":"TX-1","text":"fix: 2.388E+3 /MW\ncap: 5.16E+5 /MW","category":"page"},{"location":"README_TX_1/#","page":"TX-1","title":"TX-1","text":"##Other Alternatively for price of 1.0/W edit .csv files and replace costs with:","category":"page"},{"location":"README_TX_1/#","page":"TX-1","title":"TX-1","text":"fix: 4.776E+3 /MW  cap: 1.032E+6 /MW","category":"page"},{"location":"README_TX_1/#","page":"TX-1","title":"TX-1","text":"Assumptions for transformation:","category":"page"},{"location":"README_TX_1/#","page":"TX-1","title":"TX-1","text":"demand multiplied with 1.48\nsolar divided by 1000","category":"page"},{"location":"opt_cep_examples/#Examples-1","page":"Examples","title":"Examples","text":"","category":"section"},{"location":"opt_cep_examples/#","page":"Examples","title":"Examples","text":"Here, we present some examples.","category":"page"},{"location":"opt_cep_examples/#CO2-Limitation-1","page":"Examples","title":"CO2-Limitation","text":"","category":"section"},{"location":"opt_cep_examples/#","page":"Examples","title":"Examples","text":"using CapacityExpansion\nusing Clp\noptimizer=Clp.Optimizer #select an Optimize\nstate=\"GER_1\" #select state\nts_input_data = load_timeseries_data_provided(state; T=24)\ncep_data = load_cep_data_provided(state)\nts_clust_data = run_clust(ts_input_data;method=\"kmeans\",representation=\"centroid\",n_init=5,n_clust=5).clust_data\n# tweak the CO2 level\nco2_result = run_opt(ts_clust_data,cep_data,optimizer;descriptor=\"co2\",limit_emission=Dict{String,Number}(\"CO2/electricity\"=>50))","category":"page"},{"location":"opt_cep_examples/#Slack-variables-included-1","page":"Examples","title":"Slack variables included","text":"","category":"section"},{"location":"opt_cep_examples/#","page":"Examples","title":"Examples","text":"slack_result = run_opt(ts_clust_data,cep_data,optimizer;descriptor=\"slack\",lost_load_cost=Dict{String,Number}(\"electricity\"=>1e6), lost_emission_cost=Dict{String,Number}(\"CO2\"=>700))","category":"page"},{"location":"opt_cep_examples/#Simple-storage-1","page":"Examples","title":"Simple storage","text":"","category":"section"},{"location":"opt_cep_examples/#","page":"Examples","title":"Examples","text":"note: Note\nIn simple or intradaystorage the storage level is enforced to be the same at the beginning and end of each day. The variable 'INTRASTORAGE' is tracking the storage level within each day of the representative periods.","category":"page"},{"location":"opt_cep_examples/#","page":"Examples","title":"Examples","text":"simplestor_result = run_opt(ts_clust_data
[…]
