### 2020-08-28 #2034 @YoungFaithful
Dear @katyhuff and @labarba, sorry for the delay on my end. I now found the time to go through the paper and docs and to find any grammar and spelling mistakes left. I connected the repository with Zenodo and the DOI is `DOI: 10.5281/zenodo.4003566` Thanks for your support!
